Made fix suggested by Maarten Steenhagen
Evincing "a grasp of the distinction between a gerund and a participle", as prescribed by Dummett

"People frequently remark that they see no point in observing grammatical rules, so long as they convey their meaning. This is like saying that there is nothing wrong with using a razor blade to cut string, so long as the string is cut. By violating the rules, they make it difficult for others to express their meaning without ambiguity."